Modified all write_sha1_file calls to be asyncronous, this way not only blobs are written on multiple threads, but trees, chunks, snapshots etc + fixed a couple of free() bugs which appeared after switching to async

+++ file.h
@@ -20,7 +20,7 @@
 #define FILE_CHUNK_SIZE (10 * (1024 * 1024))
 
 int write_file(char *path, off_t size, unsigned char *sha1);
-int read_blob(unsigned char *sha1, char **out_buff, size_t *out_size);
+int read_blob(unsigned char *sha1, char **out_buff, int *out_size);
 int read_chunks_file(unsigned char *sha1, unsigned char **out_buff, int *num_chunks);
 int print_chunks_file(unsigned char *sha1);
 #endif
